+++ src/components/profile/AdminTrialManager.tsx
